requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
reportlab==3.6.12
Pillow==9.5.0
